+++ selfdrive/ui/ui.cc
@@ -367,7 +367,7 @@ void Device::updateWakefulness(const UIState &s) {
 
   // AleSato turn off screen with fog light
   SubMaster &sm = *(s.sm);
-  bool should_shutdown = Params().getBool("AleSato_ShutdownScreen") && sm["carState"].getCarState().getGenericToggle();
+  bool should_shutdown = Params().getBool("AleSato_ShutdownScreen") && !Params().getBool("ExperimentalMode") && sm["carState"].getCarState().getGenericToggle();
   setAwake((s.scene.ignition && !should_shutdown) || interactive_timeout > 0);
 }